Add files via upload
Ransomware is a type of malicious software (malware) designed with the intent to encrypt a user's files or lock them out of their computer system, rendering the data inaccessible. Perpetrators behind ransomware attacks then demand a ransom, usually in cryptocurrency, from the victim in exchange for the decryption key or to regain access to their files or system.